% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnovaTest.R
\name{AnovaTest}
\alias{AnovaTest}
\title{Function to perform Anova test}
\usage{
AnovaTest(data, VIF1, VIF3, defect1, defect2)
}
\arguments{
\item{Collect}{Chisq values of all the metrics involved in LR, Defaults to dataset, VIF1, VIF3, defect.}
}
\description{
A function that automates the process of performing Anova type 2 test for the LR models.
}
\examples{
AnovaTest()
}
\keyword{VIF,}
\keyword{data}
\keyword{defect,}
