% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyVarClus.R
\name{applyVarClus}
\alias{applyVarClus}
\title{Function for VarClus.}
\usage{
applyVarClus(dataset, sw.metrics, VarClus.threshold)
}
\arguments{
\item{Collect}{non-correlated metrics, Defaults to dataset, sw.metrics, threshold of varclus (>0.7)}
}
\description{
A function that automates the process of collecting non-correlated metrics after the application of varclus.
}
\examples{
applyVarClus()
}
\keyword{output}
\keyword{sw.metrics,}
\keyword{varclus.threshold,}
