% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyVIF.R
\name{applyVIF}
\alias{applyVIF}
\title{Function to check VIF of metrics.}
\usage{
applyVIF(dataset, sw.metrics, defect, VIF.threshold)
}
\arguments{
\item{Obtain}{metrics with high VIF value (> 5), Defaults to sw.metrics, vif threshold, dataset.}
}
\description{
Function to check Variance Inflation Factor of the non-correlated metrics, before it is trained using a model.
}
\examples{
 
applyVIF()
}
\keyword{dataset}
\keyword{sw.metrics,}
\keyword{threshold,}
\keyword{vif}
