% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollectFNFP.R
\name{CollectFNFP}
\alias{CollectFNFP}
\title{Collect False Negative and False Positive Values.}
\usage{
CollectFNFP(MSD)
}
\arguments{
\item{Obtain}{the FN and FP values through a dataframe, Defaults to MSD (dataframe containing percentage values of all FN,FP)}
}
\description{
Function to Collect the False Negative and False Positive Values from each dataset, for the purpose of knowing how many percentage of mislabelled and defective modules are present.
}
\examples{
 
CollectFNFP()
}
\keyword{FN,}
\keyword{FP}
\keyword{MSD,}
