% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installpkgs.R
\name{installpkgs}
\alias{installpkgs}
\title{Function for Packages}
\usage{
installpkgs(pkg)
}
\arguments{
\item{Want}{to install multiple packages ?, Defaults to pkg.}
}
\description{
To install and load multiple packages at once.
}
\examples{
installpkgs()
}
\keyword{pkg}
