% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_diff.R
\name{rank_diff}
\alias{rank_diff}
\title{Function to collect Rank Differences between two metrics in Type_1 and Type_3 models.}
\usage{
rank_diff(skT1, skT3, i)
}
\arguments{
\item{Collect}{Rank Difference of various metrics in T1 and T3 models.}
}
\description{
A function that automates the process of deleting the highly correlated metrics after the application of varclus.
}
\examples{
rank_diff() 
}
\keyword{rank}
\keyword{skT1,}
\keyword{skT3,}
