% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCSV.R
\name{readCSV}
\alias{readCSV}
\title{A ReadCSV Function}
\usage{
readCSV(ldata)
}
\arguments{
\item{Want}{to read multiple csv files ? Defaults to list data.}
}
\description{
This Function allows to Read Multiple CSV files into a list
}
\examples{
readCSV()
}
\keyword{ldata}
