% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeCorrelatedMetrics.R
\name{removeCorrelatedMetrics}
\alias{removeCorrelatedMetrics}
\title{Function to Remove mitigated metrics}
\usage{
removeCorrelatedMetrics(dataset, sw.metrics, VarClus.threshold)
}
\arguments{
\item{Collect}{correlated metrics, Defaults to dataset, sw.metrics, threshold of varclus (>0.7)}
}
\description{
A function that automates the process of deleting the highly correlated metrics after the application of varclus.
}
\examples{
removeCorrelatedMetrics() 
}
\keyword{output}
\keyword{sw.metrics,}
\keyword{varclus.threshold,}
